refactor: integrate container and host resource collectors
