+++ docker-compose.collector.yml
@@ -1,6 +1,6 @@
 services:
   container-data-collector:
-    image: isslab/im-container-metrics-collector
+    image: isslab/im-container-metrics-collector:latest  # 직접 빌드한 이미지로 변경
     container_name: container-data-collector
     user: root
     volumes:
@@ -16,19 +16,41 @@ services:
       - monitoring_network
     restart: always
 
-  machine-data-collector:
-    image: isslab/im-host-metrics-collector
-    container_name: machine-data-collector
-    environment:
-      - BOOTSTRAP_SERVER=${BOOTSTRAP_SERVER}
-      - KAFKA_TOPIC_HOST=${KAFKA_TOPIC_HOST}
-      - TZ=${TZ}
-    ports:
-      - "8002:8002"
-    networks:
-      - monitoring_network
-    restart: always
-
 networks:
   monitoring_network:
     driver: bridge
+
+#services:
+#  container-data-collector:
+#    image: isslab/im-container-metrics-collector
+#    container_name: container-data-collector
+#    user: root
+#    volumes:
+#      - /var/run/docker.sock:/var/run/docker.sock:ro
+#    environment:
+#      - BOOTSTRAP_SERVER=${BOOTSTRAP_SERVER}
+#      - KAFKA_TOPIC_HOST=${KAFKA_TOPIC_HOST}
+#      - KAFKA_TOPIC_CONTAINER=${KAFKA_TOPIC_CONTAINER}
+#      - TZ=${TZ}
+#    ports:
+#      - "8001:8001"
+#    networks:
+#      - monitoring_network
+#    restart: always
+#
+#  machine-data-collector:
+#    image: isslab/im-host-metrics-collector
+#    container_name: machine-data-collector
+#    environment:
+#      - BOOTSTRAP_SERVER=${BOOTSTRAP_SERVER}
+#      - KAFKA_TOPIC_HOST=${KAFKA_TOPIC_HOST}
+#      - TZ=${TZ}
+#    ports:
+#      - "8002:8002"
+#    networks:
+#      - monitoring_network
+#    restart: always
+#
+#networks:
+#  monitoring_network:
+#    driver: bridge
