+++ settings.gradle
@@ -4,6 +4,4 @@ include 'metrics-backend'
 include 'metrics-backend:producer'
 include 'metrics-backend:consumer'
 include 'metrics-backend:container-data-collector'
-include 'metrics-backend:machine-data-collector'
 include 'api-backend'
-
